indexing: started moving rule indexing to kompile
