+++ src/javasources/KTool/KTool.iml
@@ -44,6 +44,7 @@
         <SOURCES />
       </library>
     </orderEntry>
+    <orderEntry type="library" name="mockito-all" level="project" />
   </component>
 </module>
