+++ src/javasources/KTool/src/org/kframework/backend/java/indexing/IndexType.java
@@ -1,0 +1,10 @@
+package org.kframework.backend.java.indexing;
+
+/**
+ * Author: OwolabiL
+ * Date: 3/1/14
+ * Time: 12:23 PM
+ */
+public interface IndexType {
+    public void buildIndex();
+}
