+++ src/javasources/KTool/src/org/kframework/backend/java/kil/Definition.java
@@ -1,6 +1,8 @@
 package org.kframework.backend.java.kil;
 
 import com.google.common.collect.Sets;
+import org.kframework.backend.java.indexing.BasicIndex;
+import org.kframework.backend.java.indexing.IndexType;
 import org.kframework.backend.java.symbolic.Transformer;
 import org.kframework.backend.java.symbolic.Visitor;
 import org.kframework.kil.ASTNode;
@@ -41,6 +43,7 @@ public class Definition extends JavaSymbolicObject {
     private final Set<KLabelConstant> kLabels;
     private final Set<KLabelConstant> frozenKLabels;
     private final Context context;
+    private IndexType index;
 
     public Definition(Context context) {
         this.context = context;
@@ -132,4 +135,11 @@ public class Definition extends JavaSymbolicObject {
         throw new UnsupportedOperationException();
     }
 
+    public void setIndex(IndexType index) {
+        this.index = index;
+    }
+
+    public IndexType getIndex() {
+        return index;
+    }
 }
